chore: rename utils.ts to verifyStaleness.ts
